Refactoring GMVAE implementation into a tf.keras.models.Model subclass

+++ deepof/train_utils.py
@@ -436,13 +436,13 @@ def autoencoder_fitting(
 
     # Convert data to tf.data.Dataset objects
     train_dataset = (
-        tf.data.Dataset.from_tensor_slices((Xs, tuple(ys)))
+        tf.data.Dataset.from_tensor_slices((tf.cast(Xs, tf.float32), tuple(ys)))
         .batch(batch_size * strategy.num_replicas_in_sync, drop_remainder=True)
         .shuffle(buffer_size=X_train.shape[0])
         .with_options(options)
     )
     val_dataset = (
-        tf.data.Dataset.from_tensor_slices((Xvals, tuple(yvals)))
+        tf.data.Dataset.from_tensor_slices((tf.cast(Xvals, tf.float32), tuple(yvals)))
         .batch(batch_size * strategy.num_replicas_in_sync, drop_remainder=True)
         .with_options(options)
     )
